Moved child components to their own directory, working form switching
